Compilation script created.
This script takes a required argument for the validated results csv,
which is output from the validation script.

Script allows for a YAML config file to override default formatting
parameters with the '-c'/'--config' command arg pointing to the config
file.  An example with the default values is found in
data/config.yaml.  This requires pyyaml to be installed.

Script also allows for controlling the output directory, similar to the
validation script.  Default is compiled_results_YYYYMMDD.